move some common functions to common.cpp and add some functions in common
